completed ip gui
