--- src/resources/credentials.txt
@@ -1,7 +1,0 @@
-dealer,marydealer,securepass2
-player,johndealer,securepass1
-dealer,carlos1234!,carlos9999
-player,carlos1,carlos1234
-Player,carlos123456,password1234
-dealer,carlos1234,carlos1234!
-player,usernam1carlos,password
